Yatra-be/requirements.txt
flask>=3.0
numpy
Pillow
python-dotenv
requests>=2.28
waitress
